+++ rerun_py/rerun_sdk/rerun/_rerun2/datatypes/tensor_data.py
@@ -68,7 +68,7 @@ class TensorDataType(BaseExtensionType):
                                 "item",
                                 pa.struct(
                                     [
-                                        pa.field("size", pa.int64(), nullable=False, metadata={}),
+                                        pa.field("size", pa.uint64(), nullable=False, metadata={}),
                                         pa.field("name", pa.utf8(), nullable=True, metadata={}),
                                     ]
                                 ),
@@ -116,7 +116,7 @@ class TensorDataType(BaseExtensionType):
                                 ),
                                 pa.field(
                                     "I16",
-                                    pa.list_(pa.field("item", pa.int8(), nullable=False, metadata={})),
+                                    pa.list_(pa.field("item", pa.int16(), nullable=False, metadata={})),
                                     nullable=False,
                                     metadata={},
                                 ),
